refactor: make sure params to send
